requirements.txt
# Existing
groq==0.9.0
atproto==0.0.46
praw==7.7.1
feedparser==6.0.11
requests==2.31.0
python-dotenv==1.0.0

# New
click==8.1.7           # CLI enhanced
pytest==7.4.3          # Testing
pytest-cov==4.1.0      # Coverage
black==23.12.1         # Code formatting
ruff==0.1.9            # Linting
